Feature - add notes to user list
A new link next to 'Edit' allows the user to Add a note. Clicking the
link reveals a note-entry area. Clicking the link again hides the note
entry area. This note entry area is always at the top of the list of
notes for that user (if any).

Notes must be saved explicitly. A save note button commits the note to
the database.

Any number of user rows can have a note editing panel open at once, but
only one note editing panel may be opened per user entry.

Note: content from textarea is not sanitized. However it is only ever
read from and written to a textarea.

Multiple notes are listed in descending chronological order.

Notes cannot be edited or deleted after they are saved. Such an ability
to edit notes (including deleting of notes) would need to be added as a
separate feature.

Database needs to be updated to newer version that supports notes!
* For this example feature addition, I will not be creating a DB
migration. According to
https://typeorm.io/migrations#how-migrations-work in production, we
would not use 'synchronize: true' as in /server/src/data-source.ts:8,
because we would need to adjust in-use databases to have the new data
schema before writing and adding our new feature "notes" to it. But, for
brevity, I will assume that part of the assignment is out-of-scope or
that our SimpleCRM is in beta ;-)